make common superclass for remap jar tasks
